Move collect_token_balances to solana-ledger
